fix: fix cast type

+++ src/lib.rs
@@ -206,8 +206,9 @@ impl Platform for V5Platform {
                             i32::from(region.bottom_right.y - 1),
                             buf.as_ptr()
                                 .add(
-                                    region.top_left.y * Display::HORIZONTAL_RESOLUTION
-                                        + region.top_left.x,
+                                    (region.top_left.y * Display::HORIZONTAL_RESOLUTION
+                                        + region.top_left.x)
+                                        as usize,
                                 )
                                 .cast_mut(),
                             i32::from(Display::HORIZONTAL_RESOLUTION),
